Initial fixes and upkeep / cleanup after fixing merge conflicts - TODO OSC stuff
