+++ EyeTrackApp/config.py
@@ -1,6 +1,7 @@
 import json
 import os.path
 import shutil
+from typing import Union
 
 from EyeTrackApp.consts import PageType, CaptureSourceType
 from pydantic import BaseModel
