+++ EyeTrackApp/settings/modules/tracking_algorithm_module.py
@@ -20,8 +20,6 @@ class TrackingAlgorithmsValidationModel(BaseValidationModel):
     gui_HSF_radius: int
     gui_blob_maxsize: int
     gui_blob_minsize: int
-    gui_speed_coefficient: str  # GUI lib does not support doubles ;-;
-    gui_min_cutoff: str  # or floats ;-;
 
     gui_skip_autoradius: int
     gui_DADDYP: int
@@ -50,11 +48,11 @@ class TrackingAlgorithmsModule(SettingsModule):
         self.gui_BLINK = f"-BLINK{widget_id}-"
         self.gui_IBO = f"-IBO{widget_id}-"
         self.gui_HSRAC = f"-HSRAC{widget_id}-"
+        self.gui_LEAPP = f"-LEAPP{widget_id}-"
+        self.gui_LEAP = f"-LEAP{widget_id}-"
         self.gui_HSF_radius = f"-HSFRADIUS{widget_id}-"
         self.gui_blob_maxsize = f"-BLOBMAXSIZE{widget_id}-"
         self.gui_blob_minsize = f"-BLOBMINSIZE{widget_id}-"
-        self.gui_speed_coefficient = f"-SPEEDCOEFFICIENT{widget_id}-"
-        self.gui_min_cutoff = f"-MINCUTOFF{widget_id}-"
 
         self.gui_skip_autoradius = f"-SKIPAUTORADIUS{widget_id}-"
         self.gui_HSRACP = f"-HSRACP{widget_id}-"
@@ -63,18 +61,19 @@ class TrackingAlgorithmsModule(SettingsModule):
         self.gui_BLOBP = f"-BLOBP{widget_id}-"
         self.gui_thresh_add = f"-THRESHADD{widget_id}-"
 
+        self.gui_RANSACBLINK = f"-RANSACBLINK{widget_id}-"
+
         self.gui_circular_crop_left = f"-CIRCLECROPLEFT{widget_id}-"
         self.gui_circular_crop_right = f"-CIRCLECROPRIGHT{widget_id}-"
         self.gui_threshold = f"-BLOBTHRESHOLD{widget_id}-"
 
+        self.gui_HSF_radius_left = f"-HSFRADIUSLEFT{widget_id}-"
+        self.gui_HSF_radius_right = f"-HSFRADIUSRIGHT{widget_id}-"
+        self.ibo_filter_samples = f"-IBOFILTERSAMPLE{widget_id}-"
+        self.calibration_samples = f"-CALIBRATIONSAMPLES{widget_id}-"
+        self.ibo_fully_close_eye_threshold = f"-CLOSETHRESH{widget_id}-"
 
     def get_layout(self):
-        return [
-            *self.get_basic_layout(),
-            *self.get_one_euro_filter_layout(),
-        ]
-
-    def get_basic_layout(self):
         return [
             [
                 sg.Text("Tracking Algorithim Settings:", background_color="#242224"),
@@ -88,7 +87,7 @@ class TrackingAlgorithmsModule(SettingsModule):
                     tooltip="Our flagship algoritim, utilizing both HSF and RANSAC for best tracking quality and lighting resistance.",
                 ),
                 sg.Combo(
-                    ["1", "2", "3", "4", "5"],
+                    ["1", "2", "3", "4", "5", "6"],
                     default_value=self.config.gui_HSRACP,
                     key=self.gui_HSRACP,
                     background_color=BACKGROUND_COLOR,
@@ -106,7 +105,7 @@ class TrackingAlgorithmsModule(SettingsModule):
                     tooltip="HSF Is a new, lower resolution tracking algorithim that provides excelent resilancy to lighting conditions and great speed.",
                 ),
                 sg.Combo(
-                    ["1", "2", "3", "4", "5"],
+                    ["1", "2", "3", "4", "5", "6"],
                     default_value=self.config.gui_HSFP,
                     key=self.gui_HSFP,
                     background_color=BACKGROUND_COLOR,
@@ -126,7 +125,7 @@ class TrackingAlgorithmsModule(SettingsModule):
                     tooltip="DADDY Uses a Deep learning algorithm. This has a big CPU usage impact.",
                 ),
                 sg.Combo(
-                    ["1", "2", "3", "4", "5"],
+                    ["1", "2", "3", "4", "5", "6"],
                     default_value=self.config.gui_DADDYP,
                     key=self.gui_DADDYP,
                     background_color=BACKGROUND_COLOR,
@@ -144,7 +143,7 @@ class TrackingAlgorithmsModule(SettingsModule):
                     tooltip="RANSAC3D provides good tracking quality, however does not do well in bad lighting conditions.",
                 ),
                 sg.Combo(
-                    ["1", "2", "3", "4", "5"],
+                    ["1", "2", "3", "4", "5", "6"],
                     default_value=self.config.gui_RANSAC3DP,
                     key=self.gui_RANSAC3DP,
                     background_color=BACKGROUND_COLOR,
@@ -158,13 +157,30 @@ class TrackingAlgorithmsModule(SettingsModule):
             [
                 sg.Checkbox(
                     "",
+                    default=self.config.gui_LEAP,
+                    key=self.gui_LEAP,
+                    background_color=BACKGROUND_COLOR,
+                    tooltip="LEAP Uses a lightweight deep learning algorithm.",
+                ),
+                sg.Combo(['1', '2', '3', '4', '5', '6'],
+                         default_value=self.config.gui_LEAPP,
+                         key=self.gui_LEAPP,
+                         background_color=BACKGROUND_COLOR,
+                         text_color='white',
+                         button_arrow_color="black",
+                         button_background_color="#6f4ca1",
+                         tooltip="Select the priority of eyetracking algorithims.",
+                         ),
+                sg.Text("LEAP", background_color=BACKGROUND_COLOR),
+                sg.Checkbox(
+                    "",
                     default=self.config.gui_BLOB,
                     key=self.gui_BLOB,
                     background_color=BACKGROUND_COLOR,
                     tooltip="Blob tracking is the oldest and worst tracking algorithm, it provides fast, though sometimes innaccurate tracking.",
                 ),
                 sg.Combo(
-                    ["1", "2", "3", "4", "5"],
+                    ["1", "2", "3", "4", "5", "6"],
                     default_value=self.config.gui_BLOBP,
                     key=self.gui_BLOBP,
                     background_color=BACKGROUND_COLOR,
@@ -176,6 +192,15 @@ class TrackingAlgorithmsModule(SettingsModule):
                 sg.Text("Blob", background_color=BACKGROUND_COLOR),
             ],
             [
+                sg.Text("Blink Algo Settings:", background_color=BACKGROUND_COLOR)
+            ],
+            [
+                sg.Checkbox(
+                    "RANSAC Quick Blink Algo",
+                    default=self.config.gui_RANSACBLINK,
+                    key=self.gui_RANSACBLINK,
+                    background_color=BACKGROUND_COLOR,
+                ),
                 sg.Checkbox(
                     "Intensity Based Openness",
                     default=self.config.gui_IBO,
@@ -187,6 +212,27 @@ class TrackingAlgorithmsModule(SettingsModule):
                     default=self.config.gui_BLINK,
                     key=self.gui_BLINK,
                     background_color=BACKGROUND_COLOR,
+                ),
+            ],
+            [
+                sg.Text("IBO Filter Sample Size", background_color=BACKGROUND_COLOR),
+                sg.InputText(
+                    self.config.ibo_filter_samples,
+                    key=self.ibo_filter_samples,
+                    size=(0, 10),
+                ),
+                sg.Text("Calibration Samples", background_color=BACKGROUND_COLOR),
+                sg.InputText(
+                    self.config.calibration_samples,
+                    key=self.calibration_samples,
+                    size=(0, 10),
+                ),
+
+                sg.Text("IBO Close Threshold", background_color=BACKGROUND_COLOR),
+                sg.InputText(
+                    self.config.ibo_fully_close_eye_threshold,
+                    key=self.ibo_fully_close_eye_threshold,
+                    size=(0, 10),
                 ),
             ],
             [
@@ -204,6 +250,9 @@ class TrackingAlgorithmsModule(SettingsModule):
                 ),
             ],
             [
+                sg.Text("Advanced Tracking Algorithm Settings:", background_color='#242224')
+            ],
+            [
                 sg.Checkbox(
                     "HSF: Skip Auto Radius",
                     default=self.config.gui_skip_autoradius,
@@ -217,6 +266,26 @@ class TrackingAlgorithmsModule(SettingsModule):
                     default_value=self.config.gui_HSF_radius,
                     orientation="h",
                     key=self.gui_HSF_radius,
+                    background_color=BACKGROUND_COLOR,
+                    tooltip="Adjusts the radius paramater for HSF. Only adjust if you are having tracking issues.",
+                ),
+            ],
+            [
+                sg.Text("Left HSF Radius:", background_color=BACKGROUND_COLOR),
+                sg.Slider(
+                    range=(1, 50),
+                    default_value=self.config.gui_HSF_radius_left,
+                    orientation="h",
+                    key=self.gui_HSF_radius_left,
+                    background_color=BACKGROUND_COLOR,
+                    tooltip="Adjusts the radius paramater for HSF. Only adjust if you are having tracking issues.",
+                ),
+                sg.Text("Right HSF Radius:", background_color=BACKGROUND_COLOR),
+                sg.Slider(
+                    range=(1, 50),
+                    default_value=self.config.gui_HSF_radius_right,
+                    orientation="h",
+                    key=self.gui_HSF_radius_right,
                     background_color=BACKGROUND_COLOR,
                     tooltip="Adjusts the radius paramater for HSF. Only adjust if you are having tracking issues.",
                 ),
@@ -263,27 +332,6 @@ class TrackingAlgorithmsModule(SettingsModule):
                     key=self.gui_blob_maxsize,
                     background_color=BACKGROUND_COLOR,
                     tooltip="Maximum size a blob can be for blob tracking.",
-                ),
-            ],
-        ]
-
-    def get_one_euro_filter_layout(self):
-        return [
-            [
-                sg.Text("One Euro Filter Paramaters:", background_color="#242224"),
-            ],
-            [
-                sg.Text("Min Frequency Cutoff", background_color=BACKGROUND_COLOR),
-                sg.InputText(
-                    self.config.gui_min_cutoff,
-                    key=self.gui_min_cutoff,
-                    size=(0, 10),
-                ),
-                sg.Text("Speed Coefficient", background_color=BACKGROUND_COLOR),
-                sg.InputText(
-                    self.config.gui_speed_coefficient,
-                    key=self.gui_speed_coefficient,
-                    size=(0, 10),
                 ),
             ],
         ]
